% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_response.R
\name{kmer_response}
\alias{kmer_response}
\title{kmer_response}
\usage{
kmer_response(fastx, k = seq(5, 11, 2), fmt = NULL)
}
\arguments{
\item{fastx}{a fastq or fasta formatted file}

\item{k}{kmer size. One value or a sequence of values}

\item{fmt}{type of input file, fasta = f, fasta with multiple lines = m, fastq = q}
}
\value{
data.frame with results
}
\description{
k should be an odd number.
In case of fastq files the file can also be in compressted format (e.g. *.fastq.tar.gz).
This parameter will be inferred from the file name, if not set.
}
