% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_calc_k_half.R
\name{kmr_calc_k_half}
\alias{kmr_calc_k_half}
\title{kmr_calc_k_half}
\usage{
kmr_calc_k_half(resp)
}
\arguments{
\item{resp}{result from genomic response}
}
\value{
integer
}
\description{
kmr_calc_k_half
}
