% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_calc_k_limit.R
\name{kmr_calc_k_limit}
\alias{kmr_calc_k_limit}
\title{kmr_calc_k_limit}
\usage{
kmr_calc_k_limit(res, limit = 0.95)
}
\arguments{
\item{res}{a dataframe}

\item{limit}{numeric between 0 and 1}
}
\value{
list k and limit values
}
\description{
Calculates the k corresponding to a limit relative to the distinct number of 
k-mers.
}
