% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_compare.R
\name{kmr_compare}
\alias{kmr_compare}
\title{kmr_compare}
\usage{
kmr_compare(in_files, cmp, ci = 2, cx = 1e+09, cs = 255)
}
\arguments{
\item{in_files}{list of named file paths. The names are to be used in the
comparison string.}

\item{cmp}{a list containing a single named string with the kmer operations
using the file short names.}

\item{ci}{integer; min k cutoff value; default 2.}

\item{cx}{integer; max k cutoff value; default 1e9.}

\item{cs}{integer; max k count value; default 255.}
}
\value{
out_file
}
\description{
Compare k-mer sets
}
\details{
The in_files list strings may contain parameter values for each in_file after
the path.

Shorthand operator definitions:
 * intersect
 - subtract
 ~ counters_subtract
 + union

 Highest priority: *. All other operators have equal priority.

 Counter calculation modes are available for the oerators: *, ~,  +.
 They are:
 min, max, diff, sum, left, right
}
\examples{

if (interactive()) {
  fa <- system.file("testdata/phix174-pe_w_err_5k_30q.fastq.gz",
    package = "kmerize"
  )
  fm <- system.file("testdata/phix174_m-pe_w_err_5k_30q.fastq.gz",
    package = "kmerize"
  )

  k <- 9
  tmd <- tempdir()
  if (!dir.exists(tmd)) dir.create(tmd)

  out_fx <- file.path(tmd, "phx")
  out_fm <- file.path(tmd, "phm")

  kmr_count(fa, out_fx, k = k, f = "q")
  kmr_count(fm, out_fm, k = k, f = "q")

  cmp_xy <- file.path(tmd, "cmp_xy")
}
}
\references{
Kokot M, Dlugosz M, Deorowicz S. KMC 3: counting and
     manipulating
      k-mer statistics. Bioinformatics. 2017 Sep 1;33(17):2759-61.
}
\seealso{
Other kmer_help: 
\code{\link{kmr_length_fasta}()},
\code{\link{kmr_map_kmers}()},
\code{\link{kmr_position_coverage}()}
}
\author{
Reinhard Simon
}
\concept{kmer_help}
