% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_histogram.R
\name{kmr_histogram}
\alias{kmr_histogram}
\title{kmr_histogram}
\usage{
kmr_histogram(db, out = paste0("hist_", basename(db), ".txt"))
}
\arguments{
\item{db}{path to kmc3 database}

\item{out}{filename for histogram data}
}
\value{

}
\description{
kmr_histogram
}
