% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_install_kmc.R
\name{kmr_install_kmc}
\alias{kmr_install_kmc}
\title{kmr_install_kmc}
\usage{
kmr_install_kmc(download_file = NULL)
}
\arguments{
\item{download_file}{If provided, kmc3 is provided from a local archive.
User makes sure it is the appropriate version for the local OS.}
}
\value{
logical
}
\description{
Installs a local version of kmc3.
}
\examples{
if (interactive()) {
  kmr_install_kmc()
}
}
\seealso{
Other kmer_core: 
\code{\link{kmr_count}()},
\code{\link{kmr_read_parquet}()},
\code{\link{kmr_read_rds}()},
\code{\link{kmr_read_tab}()},
\code{\link{kmr_write_parquet}()},
\code{\link{kmr_write_rds}()},
\code{\link{kmr_write_tab}()}
}
\concept{kmer_core}
