% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_length_fasta.R
\name{kmr_length_fasta}
\alias{kmr_length_fasta}
\title{kmr_length_fasta}
\usage{
kmr_length_fasta(fasta = NULL)
}
\arguments{
\item{fasta}{path to a file}
}
\value{
integer length of the sequence in the file
}
\description{
Get the length of a sequnce in a fasta formatted file.
}
