% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_plot_map.R
\name{kmr_plot_map}
\alias{kmr_plot_map}
\title{kmr_plot_map}
\usage{
kmr_plot_map(kmers, snp_pos)
}
\arguments{
\item{kmers}{a GenomicRange object}

\item{snp_pos}{a position on the sequence}
}
\value{
ggbio plot
}
\description{
Create a quick overview of the kmers mapped to a region around the given
position.
}
\examples{
if (interactive()) {

  fa <- system.file("testdata/phix174.fasta", package = "kmerize")
  a <- system.file("testdata/phix174-pe_w_err_5k_30q.fastq.gz",
                   package = "kmerize")
  b <- system.file("testdata/phix174_m-pe_w_err_5k_30q.fastq.gz",
                   package = "kmerize")
  
  x <- kmr_scan_k_min(a, b, k = 9, min_kmers = 3, cleanup = FALSE)
  rr <- kmr_read_rds(x$kmer_tbl)
  
  
  kmap <- kmr_map_kmers(rr, genome = fa, "phi-X174")
  kmr_plot_map(kmap, snp_pos = 911)
}


}
\seealso{
Other kmer_help: 
\code{\link{kmr_compare}()},
\code{\link{kmr_length_fasta}()},
\code{\link{kmr_map_kmers}()},
\code{\link{kmr_plot_response}()},
\code{\link{kmr_position_coverage}()},
\code{\link{kmr_response}()},
\code{\link{kmr_scan_k_min}()}
}
\concept{kmer_help}
