% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_plot_response.R
\name{kmr_plot_response}
\alias{kmr_plot_response}
\title{kmr_plot_response}
\usage{
kmr_plot_response(res, ref_k = NULL, limit = 0.95)
}
\arguments{
\item{res}{result from genomic response scan}

\item{ref_k}{a reference k value to include}

\item{limit}{a numeric value between 0 and 1. If 0 no line will be displayed}
}
\value{
plot
}
\description{
Plots genomics response object.
}
