% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_position_coverage.R
\name{kmr_position_coverage}
\alias{kmr_position_coverage}
\title{kmr_position_coverage}
\usage{
kmr_position_coverage(
  kmers = NULL,
  pos = NULL,
  fasta = fasta,
  min_coverage = 5
)
}
\arguments{
\item{kmers}{a list object as returned from kmr_scan_k_min}

\item{pos}{a position from the reference genome}

\item{fasta}{a path to a fasta sequence file (for example a reference
chromosome)}

\item{min_coverage}{a desirable level of minimum coverage}
}
\value{
data.frame
}
\description{
Scans a set of mapped kmers if they fall around a position.
Returns a summary table for each k which includes the coverage and the start
and beginning of the range.
}
\examples{
if (interactive()) {

  # Reference genome
  fasta <- system.file("testdata/phix174.fasta", package = "kmerize")

  # Simulated NGS files from reference sequence and a modified sequence.
  # The modification is a SNP at position 911.

  a <- system.file("testdata/phix174-pe_w_err_5k_30q.fastq.gz",
    package = "kmerize"
  )
  b <- system.file("testdata/phix174_m-pe_w_err_5k_30q.fastq.gz",
    package = "kmerize"
  )

  
  kmers <- kmr_scan_k_min(a, b,
    k = seq(5, 15, 2),
    min_kmers = 1,
    cleanup = FALSE,
    ci = 0,
    cx = 1000
  )

  kmr_position_coverage(kmers, pos = 911, fasta = fasta, min_coverage = 5)
}
}
\seealso{
Other kmer_help: 
\code{\link{kmr_compare}()},
\code{\link{kmr_length_fasta}()},
\code{\link{kmr_map_kmers}()},
\code{\link{kmr_plot_map}()},
\code{\link{kmr_plot_response}()},
\code{\link{kmr_response}()},
\code{\link{kmr_scan_k_min}()}
}
\concept{kmer_help}
