% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_read_rds.R
\name{kmr_read_rds}
\alias{kmr_read_rds}
\title{kmr_read_rds}
\usage{
kmr_read_rds(afile)
}
\arguments{
\item{afile}{file name (obligatory extension .rds)}
}
\value{
data.fraee
}
\description{
Read a file in rds format.
}
