% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_read_tab.R
\name{kmr_read_tab}
\alias{kmr_read_tab}
\title{kmr_read_tab}
\usage{
kmr_read_tab(afile)
}
\arguments{
\item{afile}{file name (obligatory extension .tab)}
}
\value{
data.frame with kmers as DNAStringSet
}
\description{
Read a file in tab format.
}
