% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_response.R
\name{kmr_response}
\alias{kmr_response}
\title{kmr_response}
\usage{
kmr_response(fastx, k = c(5, 7, 9, 11), fmt = NULL, ci = 0)
}
\arguments{
\item{fastx}{a fastq or fasta formatted file}

\item{k}{kmer size. One value or a sequence of values}

\item{fmt}{type of input file, fasta = f, fasta with multiple lines = m,
fastq = q}

\item{ci}{cutoff value for min k}
}
\value{
data.frame with results
}
\description{
k should be an odd number.
In case of fastq files the file can also be in compressed format
(e.g. *.fastq.tar.gz).
This parameter will be inferred from the file name, if not set.
}
\examples{
if (interactive()) {
  fa <- system.file("testdata/phix174-pe_w_err_5k_30q.fastq.gz",
    package = "kmerize"
  )
  k <- as.integer(seq(1, 3, 2))


  res <- kmr_response(fa, k, fmt = "q")
}
}
\seealso{
Other kmer_help: 
\code{\link{kmr_compare}()},
\code{\link{kmr_length_fasta}()},
\code{\link{kmr_map_kmers}()},
\code{\link{kmr_plot_map}()},
\code{\link{kmr_plot_response}()},
\code{\link{kmr_position_coverage}()},
\code{\link{kmr_scan_k_min}()}
}
\concept{kmer_help}
