% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_saveas.R
\name{kmr_saveas}
\alias{kmr_saveas}
\title{kmr_saveas}
\usage{
kmr_saveas(db, to = db, format = c("tab", "parquet", "rds"))
}
\arguments{
\item{db}{kmer database}

\item{to}{target file name (wihout extension); default same as db name}

\item{format}{one of three: tab, parquet, rds}
}
\value{
string file path
}
\description{
Convert a kmer database to other formats.
}
