% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_scan_k_min.R
\name{kmr_scan_k_min}
\alias{kmr_scan_k_min}
\title{kmr_scan_k_min}
\usage{
kmr_scan_k_min(
  a,
  b,
  k = seq(3, 13, 2),
  ci = 2,
  cx = 100,
  min_kmers = 10,
  f = "q",
  cleanup = TRUE
)
}
\arguments{
\item{a}{sequence file path}

\item{b}{sequence file path}

\item{k}{k-mer size}

\item{ci}{minimum cutoff; default 2}

\item{cx}{maximum cutoff; default 100}

\item{min_kmers}{coverage 10}

\item{f}{format of files; q fastq; a fasta, m multiple fasta}

\item{cleanup}{delete k-mer count files}
}
\value{
list k_min identified minimal k; table of coverage for each k
}
\description{
kmr_scan_k_min
}
