% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_write_rds.R
\name{kmr_write_rds}
\alias{kmr_write_rds}
\title{kmr_write_rds}
\usage{
kmr_write_rds(db, to = db)
}
\arguments{
\item{db}{Name of kmer database}

\item{to}{Name of target file (default same as db name)}
}
\value{
logical
}
\description{
Write kmer database as rds format.
}
