% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_response.R
\name{plot_response}
\alias{plot_response}
\title{plot_response}
\usage{
plot_response(res, ref_k = NULL, show_k_opt = FALSE)
}
\arguments{
\item{res}{result from genomic response scan}

\item{ref_k}{a reference k value to include}

\item{show_k_opt}{a k derived from the}
}
\value{
plot
}
\description{
Plots genomics response object.
}
