% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_count.R
\name{kmr_count}
\alias{kmr_count}
\title{kmr_count}
\usage{
kmr_count(
  in_files,
  out_file,
  k = 11,
  m = 12,
  sm = FALSE,
  p = 9,
  f = "a",
  ci = 2,
  cx = 1e+09,
  cs = 255,
  b = FALSE,
  r = FALSE,
  n = 255,
  t = parallel::detectCores(),
  sf = 4,
  sp = 4,
  sr = 4,
  v = FALSE
)
}
\arguments{
\item{in_files}{a name of a single file or a vector of file names.}

\item{out_file}{name of the output file}

\item{k}{kmer size; default is 11.}

\item{m}{RAM memory size; default is 12.}

\item{sm}{strict memory mode; default is FALSE.}

\item{p}{signature length; default is 9.}

\item{f}{format of input files; options: a fasta, m multiple fasta, q fastq. Default is a.}

\item{ci}{minimum occurrence of kmers; default is 2.}

\item{cx}{maximum occurrence of kmers: default is 1e9.}

\item{cs}{maximum size of counter; default is 255.}

\item{b}{do not use canonical transformation of kmers. Default is FALSE.}

\item{r}{turn on RAM only. Default is FALSE.}

\item{n}{Number of bins. Default is 255.}

\item{t}{Number of threads. Default is all available number of CPUs.}

\item{sf}{Number of threads for FASTQ reading. Default is 4.}

\item{sp}{Number of threads for splitting. Default is 4.}

\item{sr}{Number of threads for second stage. Default is 4}

\item{v}{Use verbose mode. Default is FALSE.}
}
\value{
out_file
}
\description{
Count kmers.
}
\note{
Counts kmers up to a maximum of size of k = 255. Uses kmc3 in the backend on Windows, Linux, and
Mac. The parameter lists corresponds to the list of options of kmc3.
}
\examples{
if( interactive()) {

 fa <- system.file("testdata/phix174_m-pe_w_err_5k_30q.fastq.gz", package = "kmerize")
 k <- 13
 out_file <- file.path(tempdir(), "phwei11")
 kmerize:::kmr_kmc(fa, out_file, k = k, f = "q")

}
}
\references{
Kokot M, Długosz M, Deorowicz S. KMC 3: counting and manipulating k-mer statistics. Bioinformatics. 2017 Sep 1;33(17):2759-61.
}
\author{
Reinhard Simon
}
