% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_install_kmc.R
\name{kmr_install_kmc}
\alias{kmr_install_kmc}
\title{kmr_install_kmc}
\usage{
kmr_install_kmc(download_file = NULL)
}
\arguments{
\item{download_file}{ignored currently. Downloads file from original site.}
}
\value{
logical
}
\description{
Installs a local version of kmc3.
}
\examples{

if(interactive()) {
  kmr_install_kmc()
}
}
