% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_map_kmers.R
\name{kmr_map_kmers}
\alias{kmr_map_kmers}
\title{kmr_map_kmers}
\usage{
kmr_map_kmers(kmers, genome, seqname = NULL)
}
\arguments{
\item{kmers}{kmer table}

\item{genome}{a path to a fasta sequence}

\item{seqname}{name of the sequence}
}
\value{
GRange object of mapped kmers
}
\description{
kmr_map_kmers
}
\seealso{
Other kmer_help: 
\code{\link{kmr_compare}()},
\code{\link{kmr_length_fasta}()},
\code{\link{kmr_position_coverage}()}
}
\concept{kmer_help}
