% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_map_kmers.R
\name{kmr_map_kmers}
\alias{kmr_map_kmers}
\title{kmr_map_kmers}
\usage{
kmr_map_kmers(k, genome, seqname = "phix")
}
\arguments{
\item{genome}{a fasta sequence}

\item{seqname}{name of the sequence}

\item{kmers}{kmer table}
}
\value{

}
\description{
kmr_map_kmers
}
