% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_plot_map.R
\name{kmr_plot_map}
\alias{kmr_plot_map}
\title{kmr_plot_map}
\usage{
kmr_plot_map(kmers, snp_pos)
}
\arguments{
\item{kmers}{a GenomicRange object}

\item{snp_pos}{a position on the sequence}
}
\value{
ggbio plot
}
\description{
Create a quick overview of the kmers mapped to a region around the given
position.
}
