% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_plot_response.R
\name{kmr_plot_response}
\alias{kmr_plot_response}
\title{kmr_plot_response}
\usage{
kmr_plot_response(res, ref_k = NULL, limit = 0, max_y = -1)
}
\arguments{
\item{res}{result from genomic response scan}

\item{ref_k}{a reference k value to include}

\item{limit}{a numeric value between 0 and 1. If 0 no line will be displayed.
Default is 0. A recommended value is 0.95.}

\item{max_y}{a maximum y value. Default 0 indicates automatic selection.}
}
\value{
plot
}
\description{
Plots genomics response object.
}
\examples{
if (interactive()) {
  fa <- system.file("testdata/phix174-pe_w_err_5k_30q.fastq.gz",
    package = "kmerize"
  )
  fp <- system.file("testdata/phix174.fasta", 
                     package = "kmerize")
  dna <- Biostrings::readDNAStringSet(fp)

  k <- as.integer(seq(3, 17, 2))


  res <- kmr_response(fa, k, fmt = "q")
  kmr_plot_response(res, ref_k = 9, max_y = Biostrings::width(dna))
}
}
\seealso{
Other kmer_help: 
\code{\link{kmr_compare}()},
\code{\link{kmr_length_fasta}()},
\code{\link{kmr_map_kmers}()},
\code{\link{kmr_plot_map}()},
\code{\link{kmr_position_coverage}()},
\code{\link{kmr_response}()},
\code{\link{kmr_scan_k_min}()}
}
\concept{kmer_help}
