% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_read_parquet.R
\name{kmr_read_parquet}
\alias{kmr_read_parquet}
\title{kmr_read_parquet}
\usage{
kmr_read_parquet(afile)
}
\arguments{
\item{afile}{file name (obligatory extension .tab)}
}
\value{
parquet object
}
\description{
Read a file in parquet format.
}
\examples{
if (interactive()) {

  fa <- system.file("testdata/phix174_m-pe_w_err_5k_30q.fastq.gz",
  package = "kmerize")

  out_file <- file.path(tempdir(), "phx")
  k <- 7
  kc <- kmr_count(fa, out_file, k = k, f = "q")
  
  fpq <- kmr_write_parquet(out_file)
  
  dat <- kmr_read_parquet(out_file)
  dat
}

}
\seealso{
Other kmer_core: 
\code{\link{kmr_count}()},
\code{\link{kmr_install_kmc}()},
\code{\link{kmr_read_rds}()},
\code{\link{kmr_read_tab}()},
\code{\link{kmr_write_parquet}()},
\code{\link{kmr_write_rds}()},
\code{\link{kmr_write_tab}()}
}
\concept{kmer_core}
