% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_read_parquet.R
\name{kmr_read_parquet}
\alias{kmr_read_parquet}
\title{kmr_read_parquet}
\usage{
kmr_read_parquet(afile)
}
\arguments{
\item{afile}{file name (obligatory extension .tab)}
}
\value{
parquet object
}
\description{
Read a file in parquet format.
}
