% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_scan_k_min.R
\name{kmr_scan_k_min}
\alias{kmr_scan_k_min}
\title{kmr_scan_k_min}
\usage{
kmr_scan_k_min(
  a,
  b,
  k = seq(3, 13, 2),
  ci = 2,
  cx = 100,
  min_kmers = 10,
  f = "q",
  cleanup = TRUE
)
}
\arguments{
\item{a}{sequence file path}

\item{b}{sequence file path}

\item{k}{k-mer size}

\item{ci}{minimum cutoff; default 2}

\item{cx}{maximum cutoff; default 100}

\item{min_kmers}{coverage 10}

\item{f}{format of files; q fastq; a fasta, m multiple fasta}

\item{cleanup}{delete k-mer count files}
}
\value{
list k_min identified minimal k; table of coverage for each k
}
\description{
Experimental! This function may be removed.
}
\examples{
if (interactive()) {
  fa <- system.file("testdata/phix174.fasta", package = "kmerize")
  a <- system.file("testdata/phix174-pe_w_err_5k_30q.fastq.gz",
                 package = "kmerize")
  b <- system.file("testdata/phix174_m-pe_w_err_5k_30q.fastq.gz",
                 package = "kmerize")

  x <- kmr_scan_k_min(a, b, k = 9, min_kmers = 3, cleanup = FALSE)
  x
}
}
\seealso{
Other kmer_help: 
\code{\link{kmr_compare}()},
\code{\link{kmr_length_fasta}()},
\code{\link{kmr_map_kmers}()},
\code{\link{kmr_plot_map}()},
\code{\link{kmr_plot_response}()},
\code{\link{kmr_position_coverage}()},
\code{\link{kmr_response}()}
}
\concept{kmer_help}
