% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_write_parquet.R
\name{kmr_write_parquet}
\alias{kmr_write_parquet}
\title{kmr_write_parquet}
\usage{
kmr_write_parquet(db, to = db)
}
\arguments{
\item{db}{Name of kmer database}

\item{to}{Name of target file (default same as db name)}
}
\value{
logical
}
\description{
Write kmer database as parquet format.
}
\examples{

if (interactive()) {

  fa <- system.file("testdata/phix174_m-pe_w_err_5k_30q.fastq.gz",
  package = "kmerize")

  out_file <- file.path(tempdir(), "phx")
  k <- 7
  kmr_count(fa, out_file, k = k, f = "q")
  
  kmr_write_parquet(out_file)

  dat <- kmr_read_parquet(out_file)
  dat

}
}
\seealso{
Other kmer_core: 
\code{\link{kmr_count}()},
\code{\link{kmr_install_kmc}()},
\code{\link{kmr_read_parquet}()}
}
\concept{kmer_core}
